+++ tests/test_connection.py
@@ -2,12 +2,45 @@
 
 import pytest
 
+from paramiko import SFTPError
 from paramiko.ed25519key import Ed25519Key
 
 from common import conn, LOCAL, VFS
 from pathlib import Path
 from sftpretty import (CnOpts, Connection, ConnectionException,
                        HostKeysException, SSHException)
+
+
+def test_channel_exception(sftpserver):
+    '''test except blocks in _sftp_channel don't raise secondary errors'''
+    with sftpserver.serve_content(VFS):
+        with Connection(**conn(sftpserver)) as sftp:
+            sftp.close()
+        with pytest.raises(AttributeError):
+            sftp.listdir()
+
+    with sftpserver.serve_content(VFS):
+        with Connection(**conn(sftpserver)) as sftp:
+            with pytest.raises(OSError):
+                sftp.chdir('/does/not/exist')
+
+    with sftpserver.serve_content(VFS):
+        with Connection(**conn(sftpserver)) as sftp:
+            with pytest.raises(SFTPError):
+                sftp.chdir('/home/test/read.me')
+
+    with sftpserver.serve_content(VFS):
+        sftp = Connection(**conn(sftpserver))
+        sftp._transport.close()
+        with pytest.raises(SSHException):
+            sftp.listdir()
+
+    with sftpserver.serve_content(VFS):
+        with Connection(**conn(sftpserver)) as sftp:
+            sftp.timeout = 0.0001
+            with pytest.raises(TimeoutError,
+                               match='operation timed out after'):
+                sftp.listdir()
 
 
 def test_cnopts_bad_knownhosts():