+++ docs/changes.rst
@@ -1,5 +1,13 @@
-1.2.1 (current, released 2026-2-11)
+1.2.2 (current, released 2026-5-10)
 -----------------------------------
+    * adding new test for _sftp_channel exception handling.
+    * adding curve25519-sha256@libssh.org to kex list.
+    * fix for UnboundLocalError on a certain exception in _sftp_channel.
+    * removing diffie-hellman-group-exchange-sha1 from kex list per paramiko.
+    * removing ssh-rsa from public key type list per paramiko.
+
+1.2.1 (released 2026-2-11)
+--------------------------
     * adding boolean to rename for switch between posix and standard behavior.
     * change in default path behavior where cwd is set when default path is not.
     * update drivedrop to better handle all non-UNC Windows path possibilities.